Add custom hooks for fetching friends, members, removing user from room, and adding room member to friends
